Rename NFCService
