+++ FlipperZero/Core/Sources/Dependencies.swift
@@ -5,7 +5,7 @@ public func registerDependencies() {
     container.register(FlipperCentral.init, as: BluetoothCentral.self, isSingleton: true)
     container.register(FlipperCentral.init, as: BluetoothConnector.self, isSingleton: true)
     container.register(PairedFlipper.init, as: PairedDevice.self, isSingleton: true)
-    container.register(NFCService.init, as: NFCServiceProtocol.self, isSingleton: false)
+    container.register(IOSNFCService.init, as: NFCService.self, isSingleton: false)
     container.register(JSONDeviceStorage.init, as: DeviceStorage.self, isSingleton: true)
     container.register(JSONArchiveStorage.init, as: ArchiveStorage.self, isSingleton: true)
 }
