--- FlipperZero/Core/Sources/NFC/NFCServiceProtocol.swift
@@ -1,5 +1,0 @@
-public protocol NFCServiceProtocol {
-    var items: SafePublisher<[ArchiveItem]> { get }
-
-    func startReader()
-}
